+++ myodd/html/TokenBold.h
@@ -16,5 +16,5 @@ protected:
   virtual void OnPop( LOGFONT& logFont );
 
   // if this is the token we are looking for.
-  virtual bool OnIsToken(const MYODD_CHAR* lpString, int nLen ) const;
+  virtual bool OnIsToken(const MYODD_CHAR* lpString, __int64 nLen ) const;
 };
